+++ app/src/main/res/layout/activity_main.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:animateLayoutChanges="true"
     android:padding="@dimen/doub_grid_size"
-    tools:context=".MainActivity">
+    tools:context="dev.muuli.gtd.app.MainActivity">
 
     <TextView
         android:id="@+id/title"
