+++ library-android/src/main/java/dev/muuli/gtd/library/android/ToastUtil.kt
@@ -1,0 +1,12 @@
+package dev.muuli.gtd.library.android
+
+import android.content.Context
+import android.widget.Toast
+
+object ToastUtil {
+
+    fun showToast(context: Context, message: String): Toast =
+        Toast.makeText(context, message, Toast.LENGTH_SHORT).also {
+            it.show()
+        }
+}
