+++ library-compose/src/main/AndroidManifest.xml
@@ -3,7 +3,7 @@
 
     <application>
         <activity
-            android:name="com.ncorti.kotlin.template.app.ComposeActivity"
+            android:name="dev.muuli.gtd.library.compose.ComposeActivity"
             android:exported="false" />
     </application>
