+++ library-kotlin/src/main/java/com/ncorti/kotlin/app/library/FactorialCalculator.kt
@@ -1,0 +1,13 @@
+package dev.muuli.gtd.library.kotlin
+
+object FactorialCalculator {
+    private const val MAX_FACTORIAL_64BIT = 20
+
+    tailrec fun computeFactorial(input: Long, temp: Long = 1L): Long =
+        when {
+            input < 0 -> error("Factorial is not defined for negative numbers")
+            input > MAX_FACTORIAL_64BIT -> error("Only a factorial up to 20 can fit in a 64-bit Long")
+            input == 0L -> temp
+            else -> computeFactorial(input - 1, temp * input)
+        }
+}
